Split Application into smaller components
